+++ tasks/task.service.js
@@ -1,5 +1,6 @@
 const db = require('_helpers/db');
 const Task = db.Task;
+const jwtHelper = require('../_helpers/jwt');
 
 module.exports = {
     getAll,
@@ -10,6 +11,7 @@ module.exports = {
 };
 
 async function getAll() {
+    console.log(await Task.find().select('-hash'));
     return await Task.find().select('-hash');
 }
 
@@ -18,21 +20,27 @@ async function getById(id) {
 }
 
 async function create(taskParam) {
-    const task = new Task(taskParam);
-    
-    await task.save();
+    if (jwtHelper.getPermissions().includes('CREATE')) {
+        const task = new Task(taskParam);
+
+        await task.save();
+    }
 }
 
 async function update(id, taskParam) {
-    const task = await Task.findById(id);
+    if (jwtHelper.getPermissions().includes('WRITE')) {
+        const task = await Task.findById(id);
 
-    if (!task) throw 'Task not found';
+        if (!task) throw 'Task not found';
 
-    Object.assign(task, taskParam);
+        Object.assign(task, taskParam);
 
-    await task.save();
+        await task.save();
+    }
 }
 
 async function _delete(id) {
-    await Task.findByIdAndRemove(id);
+    if (jwtHelper.getPermissions().includes('DELETE')) {
+        await Task.findByIdAndRemove(id);
+    }
 }
